+++ bin/retrieve_vlocity_metadata.py
@@ -355,11 +355,11 @@ def main():
    object_map_by_vlocity_type = load_vlocity_object_map(vlocity_object_map_path)
 
    # check whether is force-dev-tool installed
-   if(not is_tool("force-dev-tool")):
+   if(platform.system() != 'Windows' and not is_tool("force-dev-tool")):
       raise SystemExit('Please install force-dev-tool first!\nFor more information look at\n"https://github.com/amtrack/force-dev-tool"')
 
    # check whether is force-dev-tool installed
-   if(not is_tool("vlocity")):
+   if(platform.system() != 'Windows' and not is_tool("vlocity")):
       raise SystemExit('Please install vlocity_build first!\nFor more information look at\n"https://github.com/vlocityinc/vlocity_build"')
 
    # loop through input (stdin)
